// server/@types/shared/models/Cas1SpaceBookingCharacteristic.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type Cas1SpaceBookingCharacteristic = 'hasEnSuite' | 'isArsonSuitable' | 'isSingle' | 'isStepFreeDesignated' | 'isSuitedForSexOffenders' | 'isWheelchairDesignated';